hdl: Optimize ALU's zr output function
This change will greatly reduce propagation delay in theory. I also thought of changing it to an Or then Not in that order, however that would technically use one more NAND gate at the core level to achieve so it would use more logic and add a tiny bit to the delay unnecessarily. This is more efficient for our hypothetical computer. :)