hdl: Fix zr output
So I was confused because I forgot about the use of the Or8Way chip, this is obvious to me now how to solve it. Do Or8Way for the least and most significant bytes separate (since we don't have an Or16Way), do an or to compare each output (meaning if there is a single 1 in either of these bytes, then the result will be one) and invert the result of that Or operation to make it so if it was 1 then it's zero, and vice versa.