hdl: ALU complete (ng works!!)
finally. it was a lot simpler than I was thinking originally. I was confused on 2's complement, I didn't realize that 2's complement still had a signed bit on the MSB. I just needed to AND to compare, this also uses a lot less gates than I was originally trying :P