Reorganize and use NOR as a primitive as well
Going to switch back to structural since I think I was wasting too much time on the behavioral at this stage in my learning. Also made the nor gate a fundamental primitive because why not. In this fantasy architecture it is made up for NOR and NAND gates fundamentally to to save a tiny bit of efficiency or something (even though that's probably not possible irl).